Refactor: 프로필 함수 리팩토링 - 팔로우 클릭 처리 #38
